+++ changelog.md
@@ -6,8 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [6.0.0] - 2020-11-13
+## [6.0.1] - 2020-11-18
+### Fixed
+- Fixed a bug when updating the balance, refund and full write-off. #279 @vaibhavpandeyvpz
+- Fixed bugs in unit tests.
 
+## [6.0.0] - 2020-11-13
 ### Added
 - Bigger and safer. There are never many tests. As always, new test cases have been added.
 - Package `brick/math` is now required.
@@ -42,8 +46,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 - The key in the currencies config will be removed in 7.x. Use "wallet.meta.currency".
 
-## [5.5.0] - 2020-10-01
+## [5.5.1] - 2020-10-18
+### Fixed
+- Fixed a bug when updating the balance, refund and full write-off. #279 @vaibhavpandeyvpz
+- Fixed bugs in unit tests.
 
+## [5.5.0] - 2020-10-01
 ### Added
 - Added brick/math dependency (Optional in version 5.x. If you already have a package installed, the library will switch to it automatically)
 - Added more php annotations, now it's easier to work with the library.
@@ -596,8 +604,10 @@ The operation is now executed in the transaction and updates the new `refund` fi
 - Exceptions: AmountInvalid, BalanceIsEmpty.
 - Models: Transfer, Transaction.
 
-[Unreleased]: https://github.com/bavix/laravel-wallet/compare/6.0.0...develop
-[6.0.0]: https://github.com/bavix/laravel-wallet/compare/5.5.0...6.0.0
+[Unreleased]: https://github.com/bavix/laravel-wallet/compare/6.0.1...develop
+[6.0.1]: https://github.com/bavix/laravel-wallet/compare/6.0.0...6.0.1
+[6.0.0]: https://github.com/bavix/laravel-wallet/compare/5.5.1...6.0.0
+[5.5.1]: https://github.com/bavix/laravel-wallet/compare/5.5.0...5.5.1
 [5.5.0]: https://github.com/bavix/laravel-wallet/compare/5.4.0...5.5.0
 [5.4.0]: https://github.com/bavix/laravel-wallet/compare/5.3.2...5.4.0
 [5.3.2]: https://github.com/bavix/laravel-wallet/compare/5.3.1...5.3.2